+++ Tests/Functional/Services/SourceStoreTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace webignition\BasilWorker\PersistenceBundle\Tests\Functional\Services;
 
 use webignition\BasilWorker\PersistenceBundle\Entity\Source;
-use webignition\BasilWorker\PersistenceBundle\Services\SourceFactory;
+use webignition\BasilWorker\PersistenceBundle\Services\Factory\SourceFactory;
 use webignition\BasilWorker\PersistenceBundle\Services\SourceStore;
 use webignition\BasilWorker\PersistenceBundle\Tests\Functional\AbstractFunctionalTest;
 
@@ -24,9 +24,9 @@ class SourceStoreTest extends AbstractFunctionalTest
             $this->sourceStore = $sourceStore;
         }
 
-        $sourceFactory = $this->container->get(SourceFactory::class);
-        self::assertInstanceOf(SourceFactory::class, $sourceFactory);
-        if ($sourceFactory instanceof SourceFactory) {
+        $sourceFactory = $this->container->get(\webignition\BasilWorker\PersistenceBundle\Services\Factory\SourceFactory::class);
+        self::assertInstanceOf(\webignition\BasilWorker\PersistenceBundle\Services\Factory\SourceFactory::class, $sourceFactory);
+        if ($sourceFactory instanceof \webignition\BasilWorker\PersistenceBundle\Services\Factory\SourceFactory) {
             $this->sourceFactory = $sourceFactory;
         }
     }
